add option --git-version and extend build configuration to support it
CMake now tests if Python is available and if it's the case calls a Python script (which itself calls
another Python script) to generate a header file containing information about the current git
revision. Those information will be printed on option --git-version.

If Python is not available a default header will be used.

If Shawn is not generated from a git repository or git is not available this will be mentioned instead
of revision information. See new Python scripts for details.

git-svn-id: https://shawn.svn.sourceforge.net/svnroot/shawn@462 3422d224-1e22-0410-b8a2-c109bc27950b